Revert "fix clippy errors with `cargo clippy --fix`"
This reverts commit dfd10facb36f12209a6e825a75102dd1e2302799.

+++ src/util.rs
@@ -77,11 +77,11 @@ impl Serialize for ReadableSize {
             write!(buffer, "{}PiB", size / PIB).unwrap();
         } else if size % TIB == 0 {
             write!(buffer, "{}TiB", size / TIB).unwrap();
-        } else if size % GIB == 0 {
+        } else if size % GIB as u64 == 0 {
             write!(buffer, "{}GiB", size / GIB).unwrap();
-        } else if size % MIB == 0 {
+        } else if size % MIB as u64 == 0 {
             write!(buffer, "{}MiB", size / MIB).unwrap();
-        } else if size % KIB == 0 {
+        } else if size % KIB as u64 == 0 {
             write!(buffer, "{}KiB", size / KIB).unwrap();
         } else {
             return serializer.serialize_u64(size);
